+++ NetworkBoard.X/ucontroller_config.h
@@ -18,8 +18,8 @@
 #ifndef ucontroller_config_H    /* Guard against multiple inclusion */
 #define ucontroller_config_H
 
-void __delay_us(int us);
-void __delay_ms(int ms);
+void __delay_us(long us);
+void __delay_ms(long ms);
 
 #endif /* _EXAMPLE_FILE_NAME_H */
